% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_GPS_analysis.R
\name{do_GPS_analysis}
\alias{do_GPS_analysis}
\title{Combines GPS prediction with given set of pSites}
\usage{
do_GPS_analysis(pSites, name = "table")
}
\arguments{
\item{pSites}{vector of phosphorylation sites}

\item{name}{if of GPS results in databases}
}
\value{

}
\description{
Combines GPS prediction with given set of pSites
}
