% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_GPS_analysis.R
\name{export_GPS_analysis}
\alias{export_GPS_analysis}
\title{Combines and exports GPS analysis tables}
\usage{
export_GPS_analysis(tables.list, output.dir)
}
\arguments{
\item{tables.list}{list of GPS analysis tables}

\item{output.dir}{}
}
\description{
Combines and exports GPS analysis tables
}
