% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_GPS_fasta.R
\name{export_GPS_fasta}
\alias{export_GPS_fasta}
\title{Subsets a given fasta file by given proteins for GPS 5.0 PTM site prediction}
\usage{
export_GPS_fasta(proteins, fasta.file, output.dir = "Data/GPS")
}
\arguments{
\item{proteins}{vector of proteins to export}

\item{fasta.file}{path to fasta file containing all sequences}

\item{output.dir}{path to folder where fasta file should be saved}
}
\value{

}
\description{
Subsets a given fasta file by given proteins for GPS 5.0 PTM site prediction
}
