% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pSites.R
\name{extract_pSites}
\alias{extract_pSites}
\title{Extract phosphosite in standard format from PhospoSitePlus data frame}
\usage{
extract_pSites(x)
}
\arguments{
\item{x}{data frame from PhosphoSitePlus}
}
\value{

}
\description{
Extract phosphosite in standard format from PhospoSitePlus data frame
}
