% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_GPS_results.R
\name{import_GPS_results}
\alias{import_GPS_results}
\title{Imports GPS 5.0 results file}
\usage{
import_GPS_results(file, name = "table")
}
\arguments{
\item{file}{path to GPS 5.0 results file}

\item{name}{(optional) identifier for imported data}
}
\value{

}
\description{
Imports GPS 5.0 results file
}
