% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_kinhub_kinase_list.R
\name{import_kinhub_kinase_list}
\alias{import_kinhub_kinase_list}
\title{Imports list of kinases from kinhub}
\usage{
import_kinhub_kinase_list(url = "http://www.kinhub.org/kinases.html")
}
\value{

}
\description{
Imports list of kinases from kinhub
}
