% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_ptmProteins.R
\name{summarize_ptmProteins}
\alias{summarize_ptmProteins}
\title{Summarize modified peptides and modification sites per proteinGroup}
\usage{
summarize_ptmProteins(
  proteinGroups,
  modificationSpecificPeptides,
  sites,
  observations,
  replicates = "groups",
  replicates.min = 2/3,
  group.by = "cell.line",
  modification = "Phospho (STY)",
  modification.abbr = "p",
  min.localization.prob = 0.75,
  add.info = c("Gene.names", "Protein.names")
)
}
\arguments{
\item{proteinGroups}{proteinGroups}

\item{modificationSpecificPeptides}{modificationSpecificPeptides}

\item{sites}{sites}

\item{observations}{observations}

\item{replicates}{observations data for biological replicates etc.}

\item{replicates.min}{min fraction to be considered identified}

\item{group.by}{observatrions data to groups by}

\item{modification}{name of modification site dataset}

\item{modification.abbr}{abbreviation to use for modification (e.g. 'p')}

\item{min.localization.prob}{localization probability threshold (e.g. 0.75)}
}
\description{
Summarize modified peptides and modification sites per proteinGroup
}
