% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_ptmProteins.R
\name{summarize_ptmProteins}
\alias{summarize_ptmProteins}
\title{Summarize modified peptides and modification sites per proteinGroup}
\usage{
summarize_ptmProteins(
  proteinGroups,
  modificationSpecificPeptides,
  sites,
  observations,
  replicates = "groups",
  replicates.min = 2/3,
  group.by = "cell.line",
  modification = "Phospho (STY)",
  modification.abbr = "p",
  min.localization.prob = 0.75,
  add.info = c("Gene.names", "Protein.names")
)
}
\arguments{
\item{min.localization.prob}{}
}
\value{

}
\description{
Summarize modified peptides and modification sites per proteinGroup
}
